% \name{circumference_circle}
% \alias{circumference_circle}
% \title{Calculate the circumference of a circle}
% \description{
  %   Calculate the circumference of a circle given its radius.
  % }
% \usage{
  %   circumference_circle(radius)
  % }
% \arguments{
  %   \item{radius}{The radius of the circle}
  % }
% \value{
  %   The circumference of the circle
  % }

% \name{area_circle}
% \alias{area_circle}
% \title{Calculate the area of a circle}
% \description{
  %   Calculate the area of a circle given its radius.
  % }
% \usage{
  %   area_circle(radius)
  % }
% \arguments{
  %   \item{radius}{The radius of the circle}
  % }
% \value{
  %   The area of the circle
  % }

% ... (document the remaining functions similarly)
